+++ src/client/components/Toc/index.tsx
@@ -4,7 +4,12 @@ import { useHistory, useLocation } from "react-router";
 import { useQuery } from "../../hooks/useQuery";
 import { parseQueryToSearch } from "../../utils/url";
 import appData from "virtual:appData";
-import { StyledToc, StyledTocItem, StyledTocItemTitle } from "./styled";
+import {
+  StyledToc,
+  StyledTocItem,
+  StyledTocItemTitle,
+  StyledTocTitle,
+} from "./styled";
 
 const { markdowns } = appData;
 
@@ -56,7 +61,12 @@ const Toc = () => {
     }
   };
 
-  return <StyledToc>{parse(curr?.toc, 0)}</StyledToc>;
+  return (
+    <StyledToc>
+      <StyledTocTitle>Table of Content</StyledTocTitle>
+      {parse(curr?.toc, 0)}
+    </StyledToc>
+  );
 };
 
 export { Toc };